+++ src/replica/tests/backup_block_service_mock.h
@@ -10,7 +10,7 @@
 #include <dsn/tool-api/task_tracker.h>
 #include <dsn/dist/block_service.h>
 #include "dist/replication/lib/replica_context.h"
-#include "dist/replication/test/replica_test/unit_test/replication_service_test_app.h"
+#include "replication_service_test_app.h"
 #include "block_service/test/block_service_mock.h"
 
 using namespace ::dsn;
