+++ src/replica/tests/clear.sh
