+++ src/replica/tests/replica_learn_test.cpp
@@ -5,7 +5,7 @@
 #include <gtest/gtest.h>
 
 #include "dist/replication/lib/replica.h"
-#include "dist/replication/test/replica_test/unit_test/mock_utils.h"
+#include "mock_utils.h"
 #include "dist/replication/lib/duplication/test/duplication_test_base.h"
 
 namespace dsn {
